+++ src/components/PhotoInput.tsx
@@ -1,6 +1,7 @@
 import { createEffect, createSignal, Switch, Match, Show, For } from "solid-js";
 import { FormComponentBase } from "../FormType";
 import Toastify from 'toastify-js'
+import { locale } from "../stores/LocaleStore";
 
 const PhotoInput: FormComponentBase = props => {
   const [label, setLabel] = createSignal('');
@@ -150,7 +151,7 @@ const PhotoInput: FormComponentBase = props => {
                   name={props.component.dataKey} />
                 <button class="bg-white text-gray-500 p-2 mr-2 rounded-full focus:outline-none h-10 w-10 hover:bg-pink-200 hover:text-pink-400 hover:border-pink-200 border-2 border-gray-300  disabled:bg-gray-200 dark:disabled:bg-gray-700 dark:disabled:text-gray-400" 
                   disabled = { disableInput() }
-                  onClick={e => { (document.getElementById("inputFile_" + props.component.dataKey) as HTMLInputElement).click() }} title={"Pilih Foto"}>
+                  onClick={e => { (document.getElementById("inputFile_" + props.component.dataKey) as HTMLInputElement).click() }} title={locale.details.language[0].uploadImage}>
                   <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="2">
                     <path stroke-linecap="round" stroke-linejoin="round" d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z" />
                   </svg>